+++ instrumentation/opentelemetry-instrumentation-psycopg2/src/opentelemetry/instrumentation/psycopg2/__init__.py
@@ -110,6 +110,7 @@ from psycopg2.extensions import (
     cursor as pg_cursor,  # pylint: disable=no-name-in-module
 )
 from psycopg2.sql import Composed  # pylint: disable=no-name-in-module
+from wrapt import ObjectProxy
 
 from opentelemetry.instrumentation import dbapi
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
@@ -158,24 +159,25 @@ class Psycopg2Instrumentor(BaseInstrumentor):
         dbapi.unwrap_connect(psycopg2, "connect")
 
     # TODO(owais): check if core dbapi can do this for all dbapi implementations e.g, pymysql and mysql
-    @staticmethod
-    def instrument_connection(connection, tracer_provider=None):
-        if not hasattr(connection, "_is_instrumented_by_opentelemetry"):
-            connection._is_instrumented_by_opentelemetry = False
-
-        if not connection._is_instrumented_by_opentelemetry:
-            setattr(
-                connection, _OTEL_CURSOR_FACTORY_KEY, connection.cursor_factory
-            )
-            connection.cursor_factory = _new_cursor_factory(
-                tracer_provider=tracer_provider
-            )
-            connection._is_instrumented_by_opentelemetry = True
-        else:
+    def instrument_connection(self, connection, tracer_provider=None):
+        if isinstance(connection, ObjectProxy):
             _logger.warning(
                 "Attempting to instrument Psycopg connection while already instrumented"
             )
-        return connection
+            return connection
+
+        db_integration = DatabaseApiIntegration(
+            __name__,
+            self._DATABASE_SYSTEM,
+            connection_attributes=self._CONNECTION_ATTRIBUTES,
+            version=__version__,
+            tracer_provider=tracer_provider,
+        )
+        db_integration.get_connection_attributes(connection)
+        db_integration.cursor_factory = _new_cursor_factory(
+            tracer_provider=tracer_provider
+        )
+        return dbapi.get_traced_connection_proxy(connection, db_integration)
 
     # TODO(owais): check if core dbapi can do this for all dbapi implementations e.g, pymysql and mysql
     @staticmethod
@@ -184,7 +186,7 @@ class Psycopg2Instrumentor(BaseInstrumentor):
             connection, _OTEL_CURSOR_FACTORY_KEY, None
         )
 
-        return connection
+        return dbapi.uninstrument_connection(connection)
 
 
 # TODO(owais): check if core dbapi can do this for all dbapi implementations e.g, pymysql and mysql